Fix token handling for SSO
The only way to get an access token when SSO is being used currently is
by logging into Planka using a browser, and retrieving the `accessToken`
cookie from the browser's cookie jar.

However, this token itself won't suffice to allow API usage, since such
a token is associated with an `httpOnlyToken`, which must also be
provided as a cookie in requests.

This PR adds handling of this `httpOnlyToken`. It's a bit hackish, but I
figure better to get this working, and focus on doing things properly in
a future version, right?

Signed-off-by: martin f. krafft <madduck@madduck.net>